complete/backend/requirements.txt
Flask
flask-cors
flask-mysqldb
mysqlclient
python-dotenv
